requirements.txt
auth0-python~=3.9
